feat: Assume package dependencies are in sibling directories, instead of relative to the current file.
This means that you don't have to manually copy dependencies, they just use package management, or manual install as siblings.

This also fixes issues I was having with node-resolve that just was flat-out not working. I changed it to just use standard node stuff.

BREAKING CHANGE: Assumes sibling directories for package dependencies, instead of relative to the file importing it.